fix: extract stores
