+++ packages/builder/src/stores/portal/index.ts
@@ -21,6 +21,7 @@ export { temporalStore } from "./temporal"
 export { navigation } from "./navigation"
 export { featureFlags } from "./featureFlags"
 export { agentsStore, selectedAgent } from "./agents"
+export { chatAppsStore } from "./chatApps"
 export { clientAppsStore } from "./clientApps"
 export { bannerStore } from "./banners"
 export { appCreationStore } from "./appCreation"
