+++ packages/frontend-core/src/api/agents.ts
@@ -1,39 +1,18 @@
 import {
   AgentToolSource,
   AgentToolSourceWithTools,
-  ChatAgentRequest,
-  ChatConversation,
-  ChatApp,
   CreateAgentRequest,
   CreateAgentResponse,
   CreateToolSourceRequest,
-  FetchAgentHistoryResponse,
   FetchAgentsResponse,
   Tool,
   UpdateAgentRequest,
   UpdateAgentResponse,
-  UpdateChatAppRequest,
 } from "@budibase/types"
 
-import { Header } from "@budibase/shared-core"
 import { BaseAPIClient } from "./types"
-import { readUIMessageStream, UIMessage, UIMessageChunk } from "ai"
-import { createSseToJsonTransformStream } from "../utils/utils"
 
 export interface AgentEndpoints {
-  agentChatStream: (
-    chat: ChatConversation,
-    workspaceId: string
-  ) => Promise<AsyncIterable<UIMessage>>
-
-  removeChat: (chatId: string) => Promise<void>
-  fetchChats: (chatAppId: string) => Promise<FetchAgentHistoryResponse>
-  fetchChatApp: (
-    agentId?: string,
-    workspaceId?: string
-  ) => Promise<ChatApp | null>
-  updateChatApp: (chatApp: UpdateChatAppRequest) => Promise<ChatApp>
-
   fetchToolSources: (agentId: string) => Promise<AgentToolSourceWithTools[]>
   fetchAvailableTools: (toolSourceType: string) => Promise<Tool[]>
   createToolSource: (
@@ -47,85 +26,7 @@ export interface AgentEndpoints {
   deleteAgent: (agentId: string) => Promise<{ deleted: true }>
 }
 
-const throwOnErrorChunk = () =>
-  new TransformStream<UIMessageChunk, UIMessageChunk>({
-    transform(chunk, controller) {
-      if (chunk.type === "error") {
-        throw new Error(chunk.errorText || "Agent action failed")
-      }
-      controller.enqueue(chunk)
-    },
-  })
-
 export const buildAgentEndpoints = (API: BaseAPIClient): AgentEndpoints => ({
-  agentChatStream: async (chat, workspaceId) => {
-    const body: ChatAgentRequest = chat
-
-    const response = await fetch("/api/agent/chat/stream", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        Accept: "application/json",
-        [Header.APP_ID]: workspaceId,
-      },
-      body: JSON.stringify(body),
-      credentials: "same-origin",
-    })
-
-    if (!response.ok) {
-      const errorBody = await response.json()
-      throw new Error(
-        errorBody.message || `HTTP error! status: ${response.status}`
-      )
-    }
-
-    if (!response.body) {
-      throw new Error("Failed to get response body")
-    }
-
-    const chunkStream = response.body
-      .pipeThrough(new TextDecoderStream())
-      .pipeThrough(createSseToJsonTransformStream<UIMessageChunk>())
-      .pipeThrough(throwOnErrorChunk())
-
-    return readUIMessageStream({
-      stream: chunkStream,
-      terminateOnError: true,
-    })
-  },
-
-  removeChat: async (chatId: string) => {
-    return await API.delete({
-      url: `/api/chat/conversations/${chatId}`,
-    })
-  },
-
-  fetchChats: async (chatAppId: string) => {
-    return await API.get({
-      url: `/api/chatapp/${chatAppId}/chats`,
-    })
-  },
-
-  fetchChatApp: async (agentId?: string, workspaceId?: string) => {
-    const query = agentId ? `?agentId=${encodeURIComponent(agentId)}` : ""
-    const headers = workspaceId
-      ? {
-          [Header.APP_ID]: workspaceId,
-        }
-      : undefined
-    return await API.get({
-      url: `/api/chatapp${query}`,
-      ...(headers && { headers }),
-    })
-  },
-
-  updateChatApp: async (chatApp: UpdateChatAppRequest) => {
-    return await API.put({
-      url: "/api/chatapp",
-      body: chatApp as any,
-    })
-  },
-
   fetchToolSources: async (agentId: string) => {
     return await API.get({
       url: `/api/agent/${agentId}/toolsource`,
